+++ src/renderers/atomics/Image/index.jsx
@@ -53,8 +53,8 @@ export default class Image extends React.Component {
 
   upImage = () => {
     this.confirmImageSize()
-    document.removeEventListener('mousemove',this.moveImage)
-    document.removeEventListener('mouseup',this.upImage)
+    document.removeEventListener('mousemove', this.moveImage)
+    document.removeEventListener('mouseup', this.upImage)
   }
 
   repareChangeSize = type => (e) => {
@@ -64,14 +64,13 @@ export default class Image extends React.Component {
     this.initialWidth = imageRect.width
     this.initialHeight = imageRect.height
     e.preventDefault()
-    let that = this
-    document.addEventListener('mousemove',that.moveImage)
-    document.addEventListener('mouseup',this.upImage)
+    document.addEventListener('mousemove', this.moveImage)
+    document.addEventListener('mouseup', this.upImage)
   }
 
   render () {
 
-    const { mediaData, language, imageControls } = this.props
+    const { mediaData, language, imageControls, imageResizable } = this.props
     const { toolbarVisible, toolbarOffset, linkEditorVisible, sizeEditorVisible, tempWidth, tempHeight } = this.state
     const blockData = this.props.block.getData()
 
@@ -169,12 +168,12 @@ export default class Image extends React.Component {
               height={height}
               {...meta}
             />
-            {toolbarVisible && <div className='bf-csize-icon right-bottom' onMouseDown={this.repareChangeSize('rightbottom')} />}
-            {toolbarVisible && <div className='bf-csize-icon left-bottom' onMouseDown={this.repareChangeSize('leftbottom')} />}
-            <div 
+            {toolbarVisible && imageResizable ? <div className='bf-csize-icon right-bottom' onMouseDown={this.repareChangeSize('rightbottom')} /> : null}
+            {toolbarVisible && imageResizable ? <div className='bf-csize-icon left-bottom' onMouseDown={this.repareChangeSize('leftbottom')} /> : null}
+            {imageResizable ? <div
               className={`bf-pre-csize ${this.reSizeType}`} 
               style={{width: `${tempWidth}px`, height:`${tempHeight}px`}}
-            />
+            /> : null}
           </div>
         </div>
         {clearFix && <div className='clearfix' style={{clear:'both',height:0,lineHeight:0,float:'none'}}></div>}
@@ -290,7 +289,6 @@ export default class Image extends React.Component {
     } else {
       return
     }
-
   }
 
   setImageLink = (e) => {
@@ -300,21 +298,38 @@ export default class Image extends React.Component {
 
   setImageLinkTarget (link_target) {
 
+    const hookReturns = this.props.hooks('set-image-link-target', link_target)(link_target)
+
+    if (hookReturns === false) {
+      return false
+    }
+
+    if (typeof hookReturns === 'string') {
+      link_target = hookReturns
+    }
+
     link_target = link_target === '_blank' ? '' : '_blank'
     this.props.editor.setValue(ContentUtils.setMediaData(this.props.editor.getValue(), this.props.entityKey, { link_target }))
     window.setImmediate(this.props.editor.forceRender)
-
   }
 
   confirmImageLink = () => {
 
-    const { tempLink: link } = this.state
+    let { tempLink: link } = this.state
+    const hookReturns = this.props.hooks('set-image-link', link)(link)
+
+    if (hookReturns === false) {
+      return false
+    }
+
+    if (typeof hookReturns === 'string') {
+      link = hookReturns
+    }
 
     if (link !== null) {
       this.props.editor.setValue(ContentUtils.setMediaData(this.props.editor.getValue(), this.props.entityKey, { link }))
       window.setImmediate(this.props.editor.forceRender)
     }
-
   }
 
   handleSizeInputKeyDown = (e) => {
@@ -336,7 +351,6 @@ export default class Image extends React.Component {
     })
 
     return
-
   }
 
   setImageHeight = ({ currentTarget }) => {
@@ -350,28 +364,58 @@ export default class Image extends React.Component {
     })
 
     return
-
   }
 
   confirmImageSize = () => {
 
     const { tempWidth: width, tempHeight: height } = this.state
-    const newImageSize = {}
+    let newImageSize = {}
 
     width !== null && (newImageSize.width = width)
     height !== null && (newImageSize.height = height)
 
+    const hookReturns = this.props.hooks('set-image-size', newImageSize)(newImageSize)
+
+    if (hookReturns === false) {
+      return false
+    }
+
+    if (hookReturns && (hookReturns.width || hookReturns.height)) {
+      newImageSize = hookReturns
+    }
+
     this.props.editor.setValue(ContentUtils.setMediaData(this.props.editor.getValue(), this.props.entityKey, newImageSize))
     window.setImmediate(this.props.editor.forceRender)
-
   }
 
   setImageFloat = (float) => {
+
+    const hookReturns = this.props.hooks('set-image-float', float)(float)
+
+    if (hookReturns === false) {
+      return false
+    }
+
+    if (typeof hookReturns === 'string') {
+      float = hookReturns
+    }
+
     this.props.editor.setValue(ContentUtils.setMediaPosition(this.props.editor.getValue(), this.props.block, { float }))
     this.unlockEditor()
   }
 
   setImageAlignment = (alignment) => {
+
+    const hookReturns = this.props.hooks('set-image-alignment', alignment)(alignment)
+
+    if (hookReturns === false) {
+      return false
+    }
+
+    if (typeof hookReturns === 'string') {
+      alignment = hookReturns
+    }
+
     this.props.editor.setValue(ContentUtils.setMediaPosition(this.props.editor.getValue(), this.props.block, { alignment }))
     this.unlockEditor()
   }
@@ -392,7 +436,6 @@ export default class Image extends React.Component {
         this.setState({ toolbarOffset: this.calcToolbarOffset() })
       })
     }
-
   }
 
   hideToolbar = (event) => {
@@ -405,7 +448,6 @@ export default class Image extends React.Component {
       this.unlockEditor()
       this.props.editor.requestFocus()
     })
-
   }
 
 }